requirements.txt
Flask
python-dotenv
langchain
openai
pinecone-client
tiktoken
pypdf
Pillow==9.1
pytesseract
install whatsapp-api-client-python